% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_autogluon.R
\name{install_autogluon}
\alias{install_autogluon}
\title{Install Autogluon and its dependencies}
\usage{
install_autogluon(python_version = "3.8.13")
}
\arguments{
\item{python_version}{By default, the version is 3.8.13.}
}
\description{
Install Autogluon and its dependencies
}
\details{
Once Autogluon has been installed, you can activate the conda
environment using `reticulate::use_condaenv("r-autogluon", conda = conda_binary(), required = TRUE)`
}
