% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_autokeras.R
\name{install_autokeras}
\alias{install_autokeras}
\title{Install AutoKeras and its dependencies}
\usage{
install_autokeras(python_version = "3.8.13")
}
\arguments{
\item{python_version}{By default, the version is 3.8.13.}
}
\description{
Install AutoKeras and its dependencies
}
\details{
Once Autokeras has been installed, you can activate the conda
environment using `use_condaenv("r-autokeras")`
}
