% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_autosklearn.R
\name{install_autosklearn}
\alias{install_autosklearn}
\title{Install Autosklearn and its dependencies}
\usage{
install_autosklearn(python_version = "3.8.13", numpy_version = "1.24")
}
\arguments{
\item{python_version}{By default, the version is 3.8.13.}

\item{numpy_version}{By default, the version is inferior to 1.24.}
}
\description{
Install Autosklearn and its dependencies
}
\details{
Once Autosklearn has been installed, you can activate the conda
environment using `reticulate::use_condaenv("r-autosklearn", conda = conda_binary(), required = TRUE)`
}
