% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_pycaret.R
\name{install_pycaret}
\alias{install_pycaret}
\title{Install PyCaret and its dependencies}
\usage{
install_pycaret(python_version = "3.8.13")
}
\arguments{
\item{python_version}{By default, the version is 3.8.13.}
}
\description{
Install PyCaret and its dependencies
}
\details{
Once Pycaret has been installed, you can activate the conda
environment using `use_condaenv("r-pycaret")`
}
