% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_tpot.R
\name{install_tpot}
\alias{install_tpot}
\title{Install TPOT and its dependencies}
\usage{
install_tpot(python_version = "3.8.13", numpy_version = "1.24")
}
\arguments{
\item{python_version}{By default, the version is 3.8.13.}

\item{numpy_version}{By default, the version is inferior to 1.24.}
}
\description{
Install TPOT and its dependencies
}
\details{
Once TPOT has been installed, you can activate the conda
environment using `reticulate::use_condaenv("r-tpot", conda = conda_binary(), required = TRUE)`
}
