% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R, R/install_miniconda.R,
%   R/use_condaenv.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{import}
\alias{install_miniconda}
\alias{use_condaenv}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{reticulate}{\code{\link[reticulate]{import}}, \code{\link[reticulate]{install_miniconda}}, \code{\link[reticulate:use_python]{use_condaenv}}}
}}

