% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NALB-package.R
\docType{package}
\name{NALB-package}
\alias{NALB}
\alias{NALB-package}
\title{NALB: An R tool embedding the most famous Python autoML libraries}
\description{
Provide an R interface to autoML libraries wrote in python.
}
\author{
\strong{Maintainer}: Alex Yahiaoui Martinez \email{yahiaoui-martinez.alex@geneseng.com}

}
\keyword{internal}
